requirements.txt
alembic==1.8.0
blueprint==3.4.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-APIBlueprint
flask-blueprint==1.3.0
Flask-DotEnv==0.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
packaging==21.3
parso==0.8.3
protobuf==4.21.1
pudb==2022.1.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.3.17
urllib3==1.26.9
urwid==2.1.2
urwid-readline==0.13
Werkzeug==2.1.2
zipp==3.8.0